+++ src/Smartstore.Core/Platform/Seo/Services/UrlService.cs
@@ -28,7 +28,7 @@ namespace Smartstore.Core.Seo
         private readonly ILanguageService _languageService;
         private readonly LocalizationSettings _localizationSettings;
         internal readonly SeoSettings _seoSettings;
-        internal readonly IReservedSlugTable _reservedSlugTable;
+        internal readonly IRouteHelper _routeHelper;
         private readonly PerformanceSettings _performanceSettings;
         private readonly SecuritySettings _securitySettings;
 
@@ -43,7 +43,7 @@ namespace Smartstore.Core.Seo
             IWorkContext workContext,
             IStoreContext storeContext,
             ILanguageService languageService,
-            IReservedSlugTable reservedSlugTable,
+            IRouteHelper routeHelper,
             LocalizationSettings localizationSettings,
             SeoSettings seoSettings,
             PerformanceSettings performanceSettings,
@@ -55,7 +55,7 @@ namespace Smartstore.Core.Seo
             _workContext = workContext;
             _storeContext = storeContext;
             _languageService = languageService;
-            _reservedSlugTable = reservedSlugTable;
+            _routeHelper = routeHelper;
             _localizationSettings = localizationSettings;
             _seoSettings = seoSettings;
             _performanceSettings = performanceSettings;
@@ -80,7 +80,7 @@ namespace Smartstore.Core.Seo
                 _workContext,
                 _storeContext,
                 _languageService,
-                _reservedSlugTable,
+                _routeHelper,
                 _localizationSettings,
                 _seoSettings,
                 _performanceSettings,
@@ -568,7 +568,7 @@ namespace Smartstore.Core.Seo
                 foundIsSelf = FoundRecordIsSelf(entity, urlRecord, languageId);
 
                 // ...and it's not in the list of reserved slugs
-                var reserved = _reservedSlugTable.IsReservedSlug(tempSlug);
+                var reserved = _routeHelper.IsReservedSlug(tempSlug);
 
                 if ((urlRecord == null || foundIsSelf) && !reserved)
                 {